update for better constructure with redux
